requirements.txt
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
anthropic==0.42.0
notion-client==2.2.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
lxml==5.3.0
